Updated the composer.json to support symfony 7.0 and still support 6.4 . (#1)
This addition adds a OR support for 7.0 of the symfony framework. 

For example; 

`symfony/config:^6.4` causes an error on `composer require` if the project has a lock of `symfony/config:^7.0` so the package can't be required into any projects that are currently set up to use `7.0`. 

This change adds it as a supported option so it can be installed by those projects.