feat: document event by combining attributes and reflection (#15)
Quite a lot of changes, also BC brakes - yet we finalise the ability to
combine many documentation strategies. Currently, we have only two of
them: attributes and reflection.

The Reflection Strategy has higher priority because it fetches the least
amount of data. Ideally, we should be able to make decisions based on
class whether to even run this strategy when Properties exists.